backend/requirements.txt
Flask
Werkzeug
python-dotenv
FLask-SQLAlchemy